Fixing quirks in CanaryWorldManager
The list of existing worlds is now updated when new worlds are added
during runtime.
The methods getAllWorlds and getExistingWorlds will now return
unmodifiable lists of the relevant data to prevent outside modifications
of the lists themselves.

The removeWorlds method is now working synchronized to avoid CMEs in some
rare situations.