Fix consistency within CanaryDoubleChest
CanaryDoubleChest is constructed with one half of the chest as the inventory, which is necessary for CanaryTileEntity to set the location, but it causes all slot getting and setting operations in the underlying CanaryChest to be limited to the range 0-26.  Resetting the inventory field to equal the InventoryLargeChest fixes the issue.