Close #104
Bed spawns and world spawns function correctly now.
If a bed was obstructed or went missing the player will
instead spawn in the overworld dimension of his current world.

If the bed is okay and not obstructed, the player will be taken to
the world his bed is in and spawn there.